Breaks up tor config tests across multiple files
Separate the general system tor config from the more granular hidden
service checks.
